+++ NEWS.md
@@ -1,5 +1,9 @@
 # patRoon 0.1.0.9000
 
+## November 2018
+* Fixed: Added timeout after restarting failed command (e.g. MetFrag CL) to prevent rare error "The requested operation cannot be performed on a file with a user-mapped section open".
+
+
 ## October 2018
 * openReport option for reportMD(). If TRUE the generated report will be opened with a web browser.
 * reportPlots option for reportMD() which collapses reportFGroups, reportChord and reportFormulaSpectra and adds control to plot Venn and UpSet diagrams.